+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - [#721](https://github.com/estruyf/vscode-front-matter/issues/721): Fix keywords regex to support unicode characters
 - [#725](https://github.com/estruyf/vscode-front-matter/issues/725): Fix for opening menu of pinned items
+- [#730](https://github.com/estruyf/vscode-front-matter/issues/730): Add debounce to the input fields
 
 ## [9.4.0] - 2023-12-12 - [Release notes](https://beta.frontmatter.codes/updates/v9.4.0)